Update Iken, add Aurora
