+++ src/en/aurorascans/build.gradle
@@ -1,0 +1,10 @@
+ext {
+    extName = 'Aurora Scans'
+    extClass = '.AuroraScans'
+    themePkg = 'iken'
+    baseUrl = "https://aurorascans.com"
+    overrideVersionCode = 0
+    isNsfw = false
+}
+
+apply from: "$rootDir/common.gradle"
