+++ src/en/aurorascans/src/eu/kanade/tachiyomi/extension/en/aurorascans/AuroraScans.kt
@@ -1,0 +1,9 @@
+package eu.kanade.tachiyomi.extension.en.aurorascans
+
+import eu.kanade.tachiyomi.multisrc.iken.Iken
+
+class AuroraScans : Iken(
+    "Aurora Scans",
+    "en",
+    "https://aurorascans.com",
+)
